feat(dashboard): add context usage contracts
Export the shared ContextUsage schema and thread it through conversation
response and message-stream payload types.
