+++ src/dashboard/server/services/conversation-service.ts
@@ -10,7 +10,7 @@
 import { readdir, readFile, stat, watch, open } from 'node:fs/promises';
 import { homedir } from 'node:os';
 import { join } from 'node:path';
-import type { ChatMessage, CompactBoundary, ProposedPlan, WorkLogEntry } from '@panctl/contracts';
+import type { ChatMessage, CompactBoundary, ContextUsage, ProposedPlan, WorkLogEntry } from '@panctl/contracts';
 import { calculateCostSync, getPricingSync, type AIProvider } from '../../../lib/cost.js';
 import { MODEL_CAPABILITIES, resolveModelIdSync } from '../../../lib/model-capabilities.js';
 import { encodeClaudeProjectDir } from '../../../lib/paths.js';
@@ -81,13 +81,6 @@ export interface ConversationActivitySummary {
   isWorking: boolean;
   /** Tool name of the most recently pending tool call, if any (e.g. "Bash", "Read"). */
   currentTool: string | null;
-}
-
-export interface ContextUsage {
-  activeBytes: number;
-  estimatedTokens: number;
-  contextWindow: number;
-  percentUsed: number;
 }
 
 // ─── CWD encoding ─────────────────────────────────────────────────────────────
